+++ egui/src/context.rs
@@ -119,7 +119,7 @@ impl CtxRef {
             let show_error = |pos: Pos2, text: String| {
                 let painter = self.debug_painter();
                 let rect = painter.error(pos, text);
-                if let Some(pointer_pos) = self.input.pointer.tooltip_pos() {
+                if let Some(pointer_pos) = self.input.pointer.hover_pos() {
                     if rect.contains(pointer_pos) {
                         painter.error(
                             rect.left_bottom() + vec2(2.0, 4.0),
@@ -179,9 +179,6 @@ impl CtxRef {
     ) -> Response {
         let hovered = hovered && enabled; // can't even hover disabled widgets
 
-        let has_kb_focus = self.memory().has_kb_focus(id);
-        let lost_kb_focus = self.memory().lost_kb_focus(id);
-
         let mut response = Response {
             ctx: self.clone(),
             layer_id,
@@ -196,14 +193,29 @@ impl CtxRef {
             drag_released: false,
             is_pointer_button_down_on: false,
             interact_pointer_pos: None,
-            has_kb_focus,
-            lost_kb_focus,
             changed: false, // must be set by the widget itself
         };
 
         if !enabled || sense == Sense::hover() || !layer_id.allow_interaction() {
             // Not interested or allowed input:
+            self.memory().surrender_kb_focus(id);
             return response;
+        }
+
+        if sense.click {
+            self.memory().interested_in_kb_focus(id);
+        }
+
+        if response.has_kb_focus() && response.clicked_elsewhere() {
+            self.memory().surrender_kb_focus(id);
+        }
+
+        if sense.click
+            && response.has_kb_focus()
+            && (self.input().key_pressed(Key::Space) || self.input().key_pressed(Key::Enter))
+        {
+            // Space/enter works like a primary click for e.g. selected buttons
+            response.clicked[PointerButton::Primary as usize] = true;
         }
 
         self.register_interaction_id(id, rect.min);
@@ -643,7 +655,7 @@ impl Context {
 
     /// If `true`, egui is currently listening on text input (e.g. typing text in a [`TextEdit`]).
     pub fn wants_keyboard_input(&self) -> bool {
-        self.memory().interaction.kb_focus_id.is_some()
+        self.memory().interaction.kb_focus.focused().is_some()
     }
 
     // ---------------------------------------------------------------------
@@ -736,6 +748,17 @@ impl Context {
         ui.label(format!(
             "Wants keyboard input: {}",
             self.wants_keyboard_input()
+        ))
+        .on_hover_text("Is egui currently listening for text input");
+        ui.label(format!(
+            "keyboard focus widget: {}",
+            self.memory()
+                .interaction
+                .kb_focus
+                .focused()
+                .as_ref()
+                .map(Id::short_debug_format)
+                .unwrap_or_default()
         ))
         .on_hover_text("Is egui currently listening for text input");
         ui.advance_cursor(16.0);