+++ egui/src/widgets/selected_label.rs
@@ -55,6 +55,10 @@ impl Widget for SelectableLabel {
         let mut desired_size = total_extra + galley.size;
         desired_size.y = desired_size.y.at_least(ui.spacing().interact_size.y);
         let (rect, response) = ui.allocate_at_least(desired_size, Sense::click());
+        if response.gained_kb_focus() {
+            ui.output()
+                .push_gained_focus_event(WidgetType::SelectableLabel, &galley.text);
+        }
 
         let text_cursor = ui
             .layout()
@@ -63,7 +67,7 @@ impl Widget for SelectableLabel {
 
         let visuals = ui.style().interact_selectable(&response, selected);
 
-        if selected || response.hovered() {
+        if selected || response.hovered() || response.has_kb_focus() {
             let rect = rect.expand(visuals.expansion);
 
             let corner_radius = 2.0;